move proto packages
fix agent.go

delete central.pb.go
